requirements.txt
async-timeout==5.0.1
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.6.0
boto3==1.35.54
botocore==1.35.54
crispy-bootstrap5==0.6
sqlparse==0.4.2
cryptography==44.0.0
distro==1.9.0
Django==4.2.17
django-allauth==65.3.1
django-bootstrap-datepicker-plus==5.0.4
django-ckeditor==6.7.2
django-cleanup==5.1.0
django-crispy-forms==1.14.0
django-debug-toolbar==4.4.6
django-filter==21.1
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-report-builder==6.4.2
django-six==1.0.5
django-tables2==2.3.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
excel-base==1.0.4
exceptiongroup==1.2.2
frozenlist==1.5.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==4.11.1
iniconfig==1.1.1
isoweek==1.3.3
jdcal==1.4.1
jiter==0.8.2
jmespath==1.0.1
Markdown==3.3.7
multidict==6.1.0
mysql-connector-python==9.0.0
mysqlclient==2.2.6
numpy==1.21.4
openai==1.61.1
openpyxl==3.1.5
packaging==21.3
pandas==1.3.4
pillow==10.2.0
pkginfo==1.4.2
plotly==5.3.1
pluggy==1.0.0
propcache==0.2.1
psycopg2-binary==2.9.9
py==1.11.0
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.4.0
pyparsing==3.0.7
pypng==0.20220715.0
pytest==7.0.1
python-dateutil==2.9.0.post0
python-decouple==3.6
pytimeparse2==1.7.1
pytz==2021.3
qrcode==7.4.2
requests==2.7.0
s3transfer==0.10.3
screen==1.0.1
sentry-sdk==1.25.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sqlite3-to-mysql==2.3.0
sqlparse==0.4.2
tabulate==0.9.0
tenacity==8.0.1
TimeConvert==3.0.13
tomli==2.0.1
tqdm==4.66.4
twilio==9.4.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
Unidecode==1.3.8
urllib3==1.26.15
xlwt==1.3.0
yarl==1.18.3
zipp==3.7.0